fix(ai-v2): cast UnitData.Attack to float in aviation estimator test
Same overload-resolution issue as AviationCombatEstimator.cs — Select(x =>
x.Attack) yields List<int>, which doesn't match the IReadOnlyList<float>
snapshot overload and falls through to the IReadOnlyList<UnitData> one.

Verified with `dotnet build Assembly-CSharp-Editor.csproj` — 0 errors, 0 warnings.

+++ Assets/Editor/AiV2EconomyRaidAirIntegrationTests.cs
@@ -106,7 +106,7 @@ namespace Game.EditorTests
 
             AviationCombatEstimator.AirStrikeEstimate snapshotSupport =
                 AviationCombatEstimator.EstimateAirStrike(
-                    aircraft.Select(x => x.Attack).ToList(), 0f, 0f, defenders,
+                    aircraft.Select(x => (float)x.Attack).ToList(), 0f, 0f, defenders,
                     AirStrikePolicy.RaidSupport(42));
 
             Assert.That(support.WipeProbability, Is.Zero);
